requirements.txt
mcp>=1.0.0
rich>=13.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
psycopg2-binary>=2.9.0
mysql-connector-python>=8.0.0
pymongo>=4.0.0
